+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 1.0a10 (unreleased)
 ^^^^^^^^^^^^^^^^^^^
 
+- Show event's start date (instead of publication date) for Event-like objects in Basic, Collection and List tiles.
+  [hvelarde]
+
 - Remove PloneFormGen's 'Form Folder' from default searchable types (fixes `#438`_).
   [djowett]